requirements.txt
certifi==2019.9.11
chardet==3.0.4
idna==2.8
python-dotenv==0.10.3
redis==3.3.11
requests==2.22.0
six==1.12.0
urllib3==1.25.6
websocket-client==0.56.0
